feat: trace sync demo
