+++ src/ingester/Cargo.toml
@@ -25,7 +25,9 @@ serde_json.workspace = true
 snafu.workspace = true
 tokio.workspace = true
 async-walkdir.workspace = true
-futures-locks = "0.7.1"
+tonic.workspace = true
+opentelemetry-proto.workspace = true
+futures-locks.workspace = true
 [[bin]]
 name = "wal-reader"
 path = "bin/wal-reader.rs"
